feat: check id when delete note

+++ webapp/controllers/note.js
@@ -8,7 +8,6 @@ router.get("/all", async ctx => {
   try {
     const uid = ctx.request.uid;
     const notes = await Note.findAll({ where: { uid } });
-
     ctx.body = notes;
   } catch (err) {
     console.log(err);
@@ -32,9 +31,9 @@ router.get("/:id", async ctx => {
 
 // Create new one
 router.post("/", async ctx => {
-  const { request } = ctx;
-  const { content } = request.body;
-  const uid = request.uid;
+  const {
+    request: { content, uid }
+  } = ctx;
   try {
     const note = await Note.create({ uid, content });
     console.log(uid);
@@ -48,13 +47,14 @@ router.post("/", async ctx => {
 router.put("/:id", async ctx => {
   const {
     request: {
+      uid,
       body: { content }
     },
     params: { id }
   } = ctx;
   try {
     console.log(content);
-    const note = await Note.update({ content }, { where: { id } });
+    const note = await Note.update({ content }, { where: { id, uid } });
     ctx.body = { row_affected: note };
   } catch (err) {
     console.log(err);
@@ -64,11 +64,12 @@ router.put("/:id", async ctx => {
 // Delete one
 router.delete("/:id", async ctx => {
   const {
+    request: { uid },
     params: { id }
   } = ctx;
 
   try {
-    const note = await Note.destroy({ where: { id } });
+    const note = await Note.destroy({ where: { id, uid } });
     ctx.body = { row_affected: note };
   } catch (err) {
     console.log(err);
